UTMS_mysqlDB/requirements.txt
pyqt5  
PyQtWebEngine 
mysql-connector
sudo apt-get install python3-pyqt5  
sudo apt-get install pyqt5-dev-tools
sudo apt-get install qttools5-dev-tools